Refactor GUI into modules and improve error handling
